SQL-2831: Make the test actually pass
