+++ src/main/java/com/mongodb/jdbc/MongoDriver.java
@@ -281,6 +281,10 @@ public class MongoDriver implements Driver {
             Enumeration<?> keys = info.keys();
             while (keys.hasMoreElements()) {
                 Object potentialKey = keys.nextElement();
+                if (potentialKey == null) {
+                    throw new SQLException(
+                            "property keys must be non-null Strings, found null");
+                }
                 String key = potentialKey.toString();
                 if (key == null) {
                     throw new SQLException(
@@ -288,6 +292,10 @@ public class MongoDriver implements Driver {
                                     + potentialKey.getClass().getName());
                 }
                 Object potentialValue = info.getProperty(key);
+                if (potentialValue == null) {
+                    throw new SQLException(
+                            "property values must be non-null Strings, found null");
+                }
                 String value = potentialValue.toString();
                 if (value == null) {
                     throw new SQLException(
